dayzoffer/requirements.txt
Django==1.9.2
django-mptt==0.8.3
Pillow==3.1.1
